(refactor) Refactor reducers to handle states for current board and all boards
